html code re-write in js
