+++ src/main.js
@@ -1,5 +1,13 @@
 import './style.css';
 
-document.querySelector('#app').innerHTML = `
-  <h1>CC first prot</h1>
-`;
+import { renderScorePage } from './score/user-score.js';
+
+function score() {
+  renderScorePage();
+}
+
+score();
+
+// document.querySelector('#app').innerHTML = `
+//   <h1>CC first prot</h1>
+// `;
